pyproject.toml
[tool.poetry]
name = "001-basic-multiagenty"
version = "0.1.0"
description = ""
authors = ["julio colomer <info@aiaccelera.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^1.2.15"
langgraph = "^1.1.9"
langchain-openai = "^1.1.16"
jupyterlab = "^4.2.4"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
langchain-core = "^1.3.0"
langchain-community = "^0.4.1"
langchain-ollama = "^1.1.0"
tavily-python = "^0.7.23"
langchain-tavily = "^0.2.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
